Move rewriter to separate file
